Move WorkAggregations into a separate file
